refactor: split oversized test files to reduce analyze:files violations
Split 3 test files (>200 lines) into 9 smaller files to reduce the
count of oversized files from 26 to 23 and ratchet CI thresholds.
