+++ .github/workflows/reusable-verify.yaml
@@ -1,8 +1,8 @@
 name: Verify
 
 env:
-  MAX_FILE_COUNT: 26
-  MAX_AVG_LENGTH: 329
+  MAX_FILE_COUNT: 23
+  MAX_AVG_LENGTH: 344
   MIN_COVERAGE: 61.8
 
 on:
